manually cherry-pick testgap test_bgp_block_loopback1
